backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.3

# Database dependencies
motor==3.7.1
pymongo==4.14.1
beanie==2.0.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose==3.5.0
bcrypt==4.0.1

# Data validation
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-core==2.33.2
email-validator==2.3.0

# File handling
python-multipart==0.0.20

# Environment management
python-dotenv==1.1.1

# RAG & LLM dependencies
openai==1.35.0
pinecone-client==5.0.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text processing & chunking
langchain==0.3.7
langchain-text-splitters==0.3.0
tiktoken==0.8.0

# HTTP client for API calls
httpx==0.27.2
aiohttp==3.10.11

# Streaming responses
sse-starlette==3.0.2

# Frontend dependencies (for Streamlit)
streamlit==1.39.0
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1

# Additional utilities
# python-magic-bin==0.4.14  # Commented out for compatibility
